lambdas_code/jeager_housekeep_lambda/requirements.txt
elasticsearch-curator-api